requirements.txt
pandas==2.1.0
openpyxl==3.1.2
numpy==1.24.3
pytest==7.4.0
pyyaml==6.0.1 